requirements.txt
asgiref==3.4.1
Django==3.2.5
django-ninja==0.13.2
mysqlclient==2.0.3
Pillow==8.3.1
pydantic==1.8.2
pytz==2021.1
sqlparse==0.4.1
typing-extensions==3.10.0.0
